Fix updating of location for user
Tight now I'm at the stage where this creates a new location entry per client reload. It looks like this is where I'm going to take this before we implement a game object which would helptying this together. We might also want to take the clients id into account if we can even get any identifiers that are unique enough.